+++ genesis.go
@@ -3,7 +3,8 @@ package poa
 // NewGenesisState creates a new genesis state with default values.
 func NewGenesisState() *GenesisState {
 	return &GenesisState{
-		Params: DefaultParams(),
+		Params:            DefaultParams(),
+		PendingValidators: []Validator{},
 	}
 }
